requirements.txt
Django==3.2.7
django-cors-headers==3.8.0
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
drf-nested-routers==0.93.3
drf-yasg==1.20.0
psycopg2-binary==2.9.1
redis==3.5.3
requests==2.26.0
PyJWT==2.2.0